fix: Color convertation alghoritm

+++ include/objects/Color.h
@@ -5,7 +5,7 @@ namespace gp
 
 struct Color
 {
-    float r, g, b, a;
+    short r, g, b, a;
 
     Color() : r(0), g(0), b(0), a(1) {} 
     Color(float r, float g, float b, float a = 1) : r(r), g(g), b(b), a(a) {}
